@@ -4,7 +4,7 @@ version: 0.3.3
 homepage: https://github.com/bazrafkan/google_place
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
