@@ -1,14 +1,14 @@
 class AddressComponent {
   /// [longName] is the full text description or name of the address component as returned by the Geocoder.
-  final String longName;
+  final String? longName;
 
   /// [shortName] is an abbreviated textual name for the address component, if available.
   /// For example, an address component for the state of Alaska may have a long_name of "Alaska"
   /// and a short_name of "AK" using the 2-letter postal abbreviation.
-  final String shortName;
+  final String? shortName;
 
   /// [types] is an array indicating the type of the address component.
-  final List<String> types;
+  final List<String>? types;
 
   AddressComponent({
     this.longName,
