@@ -4,13 +4,13 @@ import 'package:google_place/src/models/viewport.dart';
 
 class Geometry {
   /// [location] contains the geocoded latitude,longitude value for this place.
-  final Location location;
+  final Location? location;
 
   /// [viewport] contains the preferred viewport when displaying this place on a map as a LatLngBounds if it is known.
-  final Viewport viewport;
+  final Viewport? viewport;
 
   /// [bounds] contains bounds object
-  final Bounds bounds;
+  final Bounds? bounds;
 
   Geometry({this.location, this.viewport, this.bounds});
 
