@@ -1,15 +1,15 @@
 class Photo {
   /// [photoReference] a string used to identify the photo when you perform a Photo request.
-  final String photoReference;
+  final String? photoReference;
 
   /// [height] the maximum height of the image.
-  final int height;
+  final int? height;
 
   /// [width] the maximum width of the image.
-  final int width;
+  final int? width;
 
   /// [htmlAttributions] contains any required attributions. This field will always be present, but may be empty.
-  final List<String> htmlAttributions;
+  final List<String>? htmlAttributions;
 
   Photo({this.photoReference, this.height, this.width, this.htmlAttributions});
 
