@@ -3,9 +3,9 @@ import 'package:http/http.dart' as http;
 
 /// The Network Utility
 class NetworkUtility {
-  static Future<String> fetchUrl(
+  static Future<String?> fetchUrl(
     Uri uri, {
-    Map<String, String> headers,
+    Map<String, String>? headers,
   }) async {
     try {
       final response =
